astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-microsoft-azure
azure-storage-file-datalake
pandas
sqlalchemy
pyodbc
python-dotenv
pyspark==3.5.3
delta-spark==3.2.0